Repository

+++ CoachMe/CoachMeIphone/Modules/Stats/Models/Stats.swift
@@ -9,11 +9,11 @@ import Foundation
 class Stats: Codable{
     var exerciseName: String        // Название упражнения
     var workingWeight: Double       // Рабочий вес
-    var repetitions: Int            // Количество повторений
-    var sets: Int                   // Количество подходов
+    var repetitions: Int16            // Количество повторений
+    var sets: Int16                   // Количество подходов
     var date: Date                  // Дата тренировки
     
-    init(exerciseName: String, workingWeight: Double, repetitions: Int, sets: Int, date: Date) {
+    init(exerciseName: String, workingWeight: Double, repetitions: Int16, sets: Int16, date: Date) {
         self.exerciseName = exerciseName
         self.workingWeight = workingWeight
         self.repetitions = repetitions
@@ -21,7 +21,7 @@ class Stats: Codable{
         self.date = date
     }
     
-    func updateProgress(workingWeight: Double, repetitions: Int, sets: Int) {
+    func updateProgress(workingWeight: Double, repetitions: Int16, sets: Int16) {
         self.workingWeight = workingWeight
         self.repetitions = repetitions
         self.sets = sets
